fix: move Gatus fetch to API and wire map outages
Keep network calls in src/api, stop dismiss from opening the status URL, and show the outage ErrorView when the map server is down.
