+++ src/components/Map/map-screen.tsx
@@ -10,6 +10,7 @@ import NativeMapCanvas from '@/components/Map/map-canvas.native'
 import { OsmCopyright } from '@/components/Map/osm-copyright'
 import { DETAIL_HIDDEN } from '@/components/Map/sheet-detents'
 import { useMapScreenChrome } from '@/hooks/useMapScreenChrome'
+import { ServiceStatus } from '@/utils/gatus-status'
 import { LoadingState } from '@/utils/ui-utils'
 import LoadingIndicator from '../Universal/loading-indicator'
 
@@ -109,6 +110,7 @@ const MapScreen = (): React.JSX.Element => {
 				>
 					<ErrorView
 						title={t('error.map.mapLoadError')}
+						statusServices={ServiceStatus.Map}
 						onButtonPress={handleRefresh}
 					/>
 				</View>
